permissions and bug fixes
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@6302 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/asset/org.wcs.smart.asset/src/org/wcs/smart/asset/ui/views/asset/AssetEditor.java
@@ -92,6 +92,7 @@ import org.wcs.smart.SmartPlugIn;
 import org.wcs.smart.asset.AssetCoreLabelProvider;
 import org.wcs.smart.asset.AssetEvents;
 import org.wcs.smart.asset.AssetPlugIn;
+import org.wcs.smart.asset.AssetSecurityManager;
 import org.wcs.smart.asset.model.Asset;
 import org.wcs.smart.asset.model.AssetDeployment;
 import org.wcs.smart.asset.model.AssetHistoryRecord;
@@ -515,26 +516,12 @@ public class AssetEditor extends EditorPart implements MapPart {
 		panel.setLayout(new GridLayout());
 		((GridLayout)panel.getLayout()).marginWidth = 0;
 		((GridLayout)panel.getLayout()).marginHeight = 0;
-		
-		ToolBar historyToolbar = new ToolBar(panel, SWT.FLAT | SWT.HORIZONTAL);
-		historyToolbar.setLayoutData(new GridData(SWT.RIGHT, SWT.FILL, true, false));
-		
-		ToolItem deleteItem = new ToolItem(historyToolbar,SWT.PUSH);
-		deleteItem.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.DELETE_ICON));
-		deleteItem.setToolTipText("delete selected history records");
-		deleteItem.addListener(SWT.Selection, e->deleteHistoryRecords());
-		deleteItem.setEnabled(false);
-		
-		ToolItem editItem = new ToolItem(historyToolbar,SWT.PUSH);
-		editItem.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.EDIT_ICON));
-		editItem.setToolTipText("edit selected history record");
-		editItem.addListener(SWT.Selection, e->editHistoryRecord());
-		editItem.setEnabled(false);
 
-		ToolItem addItem = new ToolItem(historyToolbar,SWT.PUSH);
-		addItem.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.ADD_ICON));
-		addItem.setToolTipText("create a new history record");
-		addItem.addListener(SWT.Selection, e->addHistoryRecord());
+		ToolBar historyToolbar = null;
+		if (AssetSecurityManager.INSTANCE.canEditAssetHistory()) {
+			historyToolbar = new ToolBar(panel, SWT.FLAT | SWT.HORIZONTAL);
+			historyToolbar.setLayoutData(new GridData(SWT.RIGHT, SWT.FILL, true, false));
+		}
 		
 		tblEvents = new TableViewer(panel, SWT.BORDER | SWT.FULL_SELECTION | SWT.MULTI);
 		tblEvents.setContentProvider(ArrayContentProvider.getInstance());
@@ -603,41 +590,63 @@ public class AssetEditor extends EditorPart implements MapPart {
 				return super.compare(viewer, e1, e2);
 			}
 		});
-		Menu mnu = new Menu(tblEvents.getControl());
 		
-		MenuItem mnuAdd = new MenuItem(mnu, SWT.PUSH);
-		mnuAdd.setText("New ...");
-		mnuAdd.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.ADD_ICON));
-		mnuAdd.addListener(SWT.Selection, e->addHistoryRecord());
-		
-		MenuItem mnuEdit = new MenuItem(mnu, SWT.PUSH);
-		mnuEdit.setText("Edit ...");
-		mnuEdit.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.EDIT_ICON));
-		mnuEdit.addListener(SWT.Selection, e->editHistoryRecord());
-		
-		MenuItem mnuDelete = new MenuItem(mnu, SWT.PUSH);
-		mnuDelete.setText("Delete ...");
-		mnuDelete.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.DELETE_ICON));
-		mnuDelete.addListener(SWT.Selection, e->deleteHistoryRecords());
-		
-		mnu.addMenuListener(new MenuListener() {
-			@Override
-			public void menuShown(MenuEvent e) {
-				mnuDelete.setEnabled(!tblEvents.getSelection().isEmpty());
-				mnuEdit.setEnabled(!tblEvents.getSelection().isEmpty());
-			}
+		if (AssetSecurityManager.INSTANCE.canEditAssetHistory()) {
 			
-			@Override
-			public void menuHidden(MenuEvent e) {}
-		});
-		tblEvents.addSelectionChangedListener(new ISelectionChangedListener() {
-			@Override
-			public void selectionChanged(SelectionChangedEvent event) {
-				deleteItem.setEnabled(!tblEvents.getSelection().isEmpty());
-				editItem.setEnabled(!tblEvents.getSelection().isEmpty());
-			}
-		});
-		tblEvents.getControl().setMenu(mnu);
+			ToolItem deleteItem = new ToolItem(historyToolbar,SWT.PUSH);
+			deleteItem.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.DELETE_ICON));
+			deleteItem.setToolTipText("delete selected history records");
+			deleteItem.addListener(SWT.Selection, e->deleteHistoryRecords());
+			deleteItem.setEnabled(false);
+				
+			ToolItem editItem = new ToolItem(historyToolbar,SWT.PUSH);
+			editItem.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.EDIT_ICON));
+			editItem.setToolTipText("edit selected history record");
+			editItem.addListener(SWT.Selection, e->editHistoryRecord());
+			editItem.setEnabled(false);
+		
+			ToolItem addItem = new ToolItem(historyToolbar,SWT.PUSH);
+			addItem.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.ADD_ICON));
+			addItem.setToolTipText("create a new history record");
+			addItem.addListener(SWT.Selection, e->addHistoryRecord());
+			
+			tblEvents.addSelectionChangedListener(new ISelectionChangedListener() {
+				@Override
+				public void selectionChanged(SelectionChangedEvent event) {
+					deleteItem.setEnabled(!tblEvents.getSelection().isEmpty());
+					editItem.setEnabled(!tblEvents.getSelection().isEmpty());
+				}
+			});
+		
+			Menu mnu = new Menu(tblEvents.getControl());
+			
+			MenuItem mnuAdd = new MenuItem(mnu, SWT.PUSH);
+			mnuAdd.setText("New ...");
+			mnuAdd.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.ADD_ICON));
+			mnuAdd.addListener(SWT.Selection, e->addHistoryRecord());
+			
+			MenuItem mnuEdit = new MenuItem(mnu, SWT.PUSH);
+			mnuEdit.setText("Edit ...");
+			mnuEdit.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.EDIT_ICON));
+			mnuEdit.addListener(SWT.Selection, e->editHistoryRecord());
+			
+			MenuItem mnuDelete = new MenuItem(mnu, SWT.PUSH);
+			mnuDelete.setText("Delete ...");
+			mnuDelete.setImage(SmartPlugIn.getDefault().getImageRegistry().get(SmartPlugIn.DELETE_ICON));
+			mnuDelete.addListener(SWT.Selection, e->deleteHistoryRecords());
+			
+			mnu.addMenuListener(new MenuListener() {
+				@Override
+				public void menuShown(MenuEvent e) {
+					mnuDelete.setEnabled(!tblEvents.getSelection().isEmpty());
+					mnuEdit.setEnabled(!tblEvents.getSelection().isEmpty());
+				}
+				
+				@Override
+				public void menuHidden(MenuEvent e) {}
+			});
+			tblEvents.getControl().setMenu(mnu);
+		}
 		
 		
 		initializeEventsPanel(asset);
